% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.account.changes}
\alias{oanda.account.changes}
\title{Account changes}
\usage{
oanda.account.changes(account, transactionId)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{transactionId}{The last transaction id for the account.}
}
\value{
A list
}
\description{
The function is used to poll an Account for its current state and changes since a specified TransactionID.
}
\examples{
oanda.account.changes('101-001-3704066-001',4567)
}
