% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.account.instruments.data.frame}
\alias{oanda.account.instruments.data.frame}
\title{Accounts tradable instruments}
\usage{
oanda.account.instruments.data.frame(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
A data frame of instruments and details
}
\description{
The function returns a data frame of tradeable instruments for the given Account. The list of tradeable instruments is dependent
on the regulatory division that the Account is located in, thus should be the same for all Accounts owned by a single user.
}
\examples{
oanda.account.instruments.data.frame('101-001-3704066-001')
}
