% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.account.instruments}
\alias{oanda.account.instruments}
\title{Accounts tradable instruments}
\usage{
oanda.account.instruments(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
A list of instruments and details
}
\description{
The function returns a list of tradeable instruments for the given Account. The list of tradeable instruments is dependent
on the regulatory division that the Account is located in, thus should be the same for all Accounts owned by a single user.
}
\examples{
oanda.account.instruments('101-001-3704066-001')
}
