% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.account.summary}
\alias{oanda.account.summary}
\title{Account summary}
\usage{
oanda.account.summary(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
A vector of account IDs
}
\description{
This function brings back summary such as unrealized PnL, account balance, account position, margin position,
and as well as reference data as such account creation date and risk management settings for the account.
}
\examples{
oanda.account.summary('101-001-3704066-001')
}
