% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.accounts}
\alias{oanda.accounts}
\title{Accounts}
\usage{
oanda.accounts()
}
\description{
Brings backs all of the accounts and thier respective IDs associated with the token
}
\examples{
oanda.accounts()
}
