% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.init}
\alias{oanda.init}
\title{Initialize and authenticate account}
\usage{
oanda.init(accountLevel, token)
}
\arguments{
\item{accountLevel}{The base url that points to either the practice or the production account}

\item{token}{The token generated on OANDA's website}
}
\description{
Users of this package must run this function first in order for rest of the package to work. This prevents the need to
set url and authentication token redundantly for each function.
}
\examples{
oanda.init("practice",'44667a4c1a94db7f245ea21fe2f82341-04efe2ed548b91e6bb2d7d7707668d8a')
}
