% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.instrument.orderbook}
\alias{oanda.instrument.orderbook}
\title{Instrument orderbook}
\usage{
oanda.instrument.orderbook(instrument)
}
\arguments{
\item{instrument}{The name of the instrument such as EUR_USD}
}
\value{
A named list that includes the orderbook dataframe
}
\description{
This function will return the orderbook for an instrument. Visit http://developer.oanda.com/rest-live-v20/instrument-ep/ or www.oanda.com for details
}
\examples{
oanda.instrument.orderbook("EUR_USD")
}
