% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.instrument.positionbook}
\alias{oanda.instrument.positionbook}
\title{Instrument positionbook}
\usage{
oanda.instrument.positionbook(instrument)
}
\arguments{
\item{instrument}{The name of the instrument such as EUR_USD}
}
\value{
A named list that includes the positionbook dataframe
}
\description{
This function will return the positionbook for an instrument. Visit http://developer.oanda.com/rest-live-v20/instrument-ep/ or www.oanda.com for details.
}
\examples{
oanda.instrument.positionbook("EUR_USD")
}
