% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.instrument.pricing}
\alias{oanda.instrument.pricing}
\title{Instrument Pricing}
\usage{
oanda.instrument.pricing(account, instruments)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instruments}{The name of the instrument(s)}
}
\value{
A named R list of instruments latest pricing data
}
\description{
This function will get the pricing information for a specified list of Instruments within an Account. Visit http://developer.oanda.com/rest-live-v20/instrument-ep/ or www.oanda.com for details.
}
\examples{
oanda.instrument.pricing('101-001-3704066-001','EUR_USD')
oanda.instrument.pricing('101-001-3704066-001','EUR_USD,USD_CAD')
}
