% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.instrument.stream}
\alias{oanda.instrument.stream}
\title{Instrument price stream}
\usage{
oanda.instrument.stream(account, instrument, func)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{func}{A custom function}

\item{instruments}{The name of the instrument(s)}
}
\value{
This is a void function therefore does not return any value. Without a function the default behavior of this function is to print the stream to console.
}
\description{
This endpoint is served by the streaming URLs. Get a stream of Account Prices starting from when the request is made.
This pricing stream does not include every single price created for the Account, but instead will provide at most 4 prices per second (every 250 milliseconds) for each instrument being requested.
If more than one price is created for an instrument during the 250 millisecond window, only the price in effect at the end of the window is sent. This means that during periods of rapid price movement, subscribers to this stream will not be sent every price.
Pricing windows for different connections to the price stream are not all aligned in the same way (i.e. they are not all aligned to the top of the second). This means that during periods of rapid price movement, different subscribers may observe different prices depending on their alignment.
}
\examples{
oanda.instrument.stream('101-001-3704066-001','EUR_USD')
oanda.instrument.stream('101-001-3704066-001','EUR_USD',function(x){
   #do something at each stream
   print(x) #or simply print it
})
}
