% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.instruments.bid}
\alias{oanda.instruments.bid}
\title{Instrument bid candlestick}
\source{
\url{http://developer.oanda.com/rest-live-v20/instrument-ep/}
}
\usage{
oanda.instruments.bid(
  instrument,
  granularity,
  from,
  to,
  dailyAlignment,
  weeklyAlignment,
  alignmentTimezone
)
}
\arguments{
\item{instrument}{The name of the instrument such as EUR_USD}

\item{granularity}{Frequency of the candlestick such as S5,S10,S15, etc. Please visit www.oanda.com for more details.}

\item{from}{The start of the time range in unix timestamp to fetch candlesticks for. Optional if 'to' is set.}

\item{to}{The end of the time range in unix timestamp to fetch candlesticks for. Optional if 'from' is set.}

\item{dailyAlignment}{The hour of the day (in the specified timezone) to use for granularities that have daily alignments. \link{default=17, minimum=0, maximum=23}}

\item{weeklyAlignment}{The day of the week used for granularities that have weekly alignment. \link{default=Friday}}

\item{alignmentTimezone}{The timezone to use for the dailyAlignment parameter. Candlesticks with daily alignment will be aligned to the dailyAlignment hour within the alignmentTimezone. Note that the returned times will still be represented in UTC. \link{default=America/New_York}}
}
\description{
This function will return open,high,low and closing prices from the bid side of the market. The return time series are represented in UTC no matter the time zone.
The following settings are set either by default or with purpose to reduce too many variations of one function. \cr
#price = B \cr
#count = 5000 \cr
#smooth = False \cr
#includeFirst = True //implicitly set by oanda \cr
#dailyAlignment = 17 //Optional parameter otherwise oanda sets it to 17 \cr
#weeklyAlignment = Friday //Optional parameter otherwise oanda sets it to Friday \cr
#alignmentTimezone = America/New_York //Optional parameter otherwise oanda sets it to America/New_York \cr
}
\examples{
oanda.instruments.bid("EUR_USD","M1",from = 1608064200) #2020-12-15 03:30 PM
oanda.instruments.bid("EUR_USD","M1",from = "1608064200") #2020-12-15 03:30 PM
oanda.instruments.bid("USD_CAD","S5",from="2020-12-15T15:30:00")
}
