% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.byId}
\alias{oanda.orders.byId}
\title{Pending Orders by ID}
\usage{
oanda.orders.byId(account, id)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{id}{The order ID.}
}
\value{
R list object detailing the response from the server.
}
\description{
Query all pending orders in an account by an order ID.
}
\examples{
oanda.orders.byId('101-001-3704066-001',"6985")
}
