% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.byInstrument}
\alias{oanda.orders.byInstrument}
\title{Pending Orders by Instrument}
\usage{
oanda.orders.byInstrument(account, instrument)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instrument}{The instrument to filter the requested orders by.}
}
\value{
R list object detailing the response from the server.
}
\description{
Query all pending orders in an account by instrument.
}
\examples{
oanda.orders.byInstrument('101-001-3704066-001',"USD_MXN")
}
