% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.cancelById}
\alias{oanda.orders.cancelById}
\title{Cancel Order by ID}
\usage{
oanda.orders.cancelById(account, id)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{id}{The order ID.}
}
\value{
R list object detailing the response from the server.
}
\description{
Cancel a pending order in an account.
}
\examples{
oanda.orders.cancelById('101-001-3704066-001',"6985")
}
