% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createClientExtensions}
\alias{oanda.orders.createClientExtensions}
\title{Create client extensions}
\usage{
oanda.orders.createClientExtensions(comment, tag, id)
}
\arguments{
\item{comment}{A comment associated with the Order/Trade}

\item{tag}{A tag associated with the Order/Trade}

\item{id}{The Client ID of the Order/Trade}
}
\value{
A JSON object of ClientExtensions
}
\description{
A ClientExtensions object allows a client to attach a clientID, tag and comment to Orders and Trades in their Account.
Do not set, modify, or delete this field if your account is associated with MT4.
}
\examples{
oanda.orders.createClientExtensions("A Comment","strategy_9","my_order_10000")
}
