% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createLimitOrder}
\alias{oanda.orders.createLimitOrder}
\title{Create a limit order}
\usage{
oanda.orders.createLimitOrder(
  account,
  instrument,
  units,
  price,
  clientExtensions
)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instrument}{The Market Order’s Instrument.}

\item{units}{The quantity requested to be filled by the Market Order. A positive number of units results in a long Order, and a negative number of units results in a short Order.}

\item{price}{he price threshold specified for the Limit Order. The Limit Order will only be filled by a market price that is equal to or better than this price.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
R list object detailing the response from the server.
}
\description{
This function initiates a limit order
}
\examples{
oanda.orders.createLimitOrder('101-001-3704066-001',"USD_HKD",10000,7.75405) 
#long 10000 limit order of 7.75405 of USD/HKD
info <- oanda.orders.createClientExtensions("A comment","strategy_9","my_order_10000")
oanda.orders.createLimitOrder('101-001-3704066-001',"USD_HKD",10000,7.75405, info)
#long 10000 limit order of 7.75405 of USD/HKD with client information
}
