% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createMarketIfTouchedOrder}
\alias{oanda.orders.createMarketIfTouchedOrder}
\title{Create a market if touch order}
\usage{
oanda.orders.createMarketIfTouchedOrder(
  account,
  instrument,
  units,
  price,
  priceBound,
  clientExtensions
)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instrument}{The Market Order’s Instrument.}

\item{units}{The quantity requested to be filled by the Market Order. A positive number of units results in a long Order, and a negative number of units results in a short Order.}

\item{price}{he price threshold specified for the Limit Order. The Limit Order will only be filled by a market price that is equal to or better than this price.}

\item{priceBound}{The worst price that the client is willing to have the Market Order filled at. This is optional.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
R list object detailing the response from the server.
}
\description{
This function initiates market order once a price touches a point
}
\examples{
oanda.orders.createMarketIfTouchedOrder('101-001-3704066-001', "USD_SGD", 10000, 1.32821)
#long 10000 market if touch at 1.32821 of USD/SGD
oanda.orders.createMarketIfTouchedOrder('101-001-3704066-001', "USD_SGD", 10000, 1.32821, 1.42821)
#long 10000 of USD/SGD market if touch order at 1.32821 with slippage/gap protection at 1.42821
info <- oanda.orders.createMarketIfTouchedOrder("A comment","strategy_9","my_order_10000")
oanda.orders.createStopOrder('101-001-3704066-001', "USD_SGD", 10000, 1.32821, 1.42821, info)

}
