% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createMarketOrder}
\alias{oanda.orders.createMarketOrder}
\title{Create a market order}
\usage{
oanda.orders.createMarketOrder(
  account,
  instrument,
  units,
  priceBound,
  clientExtensions
)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instrument}{The Market Order’s Instrument.}

\item{units}{The quantity requested to be filled by the Market Order. A positive number of units results in a long Order, and a negative number of units results in a short Order.}

\item{priceBound}{The worst price that the client is willing to have the Market Order filled at. This is optional.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
R list object detailing the response from the server.
}
\description{
This function initiates a market order
}
\examples{
oanda.orders.createMarketOrder('101-001-3704066-001',"USD_MXN",10000) #long 10000 USD/MXN
oanda.orders.createMarketOrder('101-001-3704066-001',"USD_MXN",-10000) #short 10000 USD/MXN
oanda.orders.createMarketOrder('101-001-3704066-001',"USD_MXN",10000,priceBound=19.86263)
#long 10000 USD/MXN with a worst pirce of 19.86263 else kill the order
info <- oanda.orders.createClientExtensions("A comment","strategy_9","my_order_10000")
oanda.orders.createMarketOrder('101-001-3704066-001',"USD_MXN",10000,priceBound=19.86263,info)
#long 10000 USD/MXN with a worst pirce of 19.86263 else kill the order
}
