% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createStopLossDetails}
\alias{oanda.orders.createStopLossDetails}
\title{Create Stop Loss details}
\usage{
oanda.orders.createStopLossDetails(price, distance, clientExtensions)

oanda.orders.createStopLossDetails(price, distance, clientExtensions)
}
\arguments{
\item{price}{The price that the Take Profit Order will be triggered at.}

\item{distance}{Specifies the distance (in price units) from the Trade’s open price to use as the Stop Loss Order price. Only one of the distance and price fields may be specified.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
A JSON object of StopLossDetails

A JSON object of StopLossDetails.
}
\description{
StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a client. This may happen when an Order is filled that opens
a Trade requiring a Stop Loss, or when a Trade’s dependent Stop Loss Order is modified directly through the Trade.

StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a client. This may happen when an Order is filled that opens
a Trade requiring a Stop Loss, or when a Trade’s dependent Stop Loss Order is modified directly through the Trade.
}
\examples{
oanda.orders.createStopLossDetails(price = 1.2345)
info <- oanda.orders.createClientExtensions("A comment","strategy_9","my_order_10000")
oanda.orders.createStopLossDetails(1.2345, info)
oanda.orders.createStopLossDetails(distance = 0.0100)
oanda.orders.createStopLossDetails(distance = 0.0100,info)
oanda.orders.createStopLossDetails(price = 1.2345)
info <- oanda.orders.createClientExtensions("A comment","strategy_9","my_order_10000")
oanda.orders.createStopLossDetails(1.2345, info)
oanda.orders.createStopLossDetails(distance = 0.0100)
oanda.orders.createStopLossDetails(distance = 0.0100, info)
}
