% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createTakeProfitDetails}
\alias{oanda.orders.createTakeProfitDetails}
\title{Create Take profit details}
\usage{
oanda.orders.createTakeProfitDetails(price, clientExtensions)
}
\arguments{
\item{price}{The price that the Take Profit Order will be triggered at}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional}
}
\value{
A JSON object of TakeProfitDetails
}
\description{
TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of a client. This may happen when an Order is filled
that opens a Trade requiring a Take Profit, or when a Trade’s dependent. Take Profit Order is modified directly through the Trade.
}
\examples{
oanda.orders.createTakeProfitDetails(1.2345)
info <- oanda.orders.createClientExtensions("100 pips away from trade open","strategy_9","my_order_10000")
oanda.orders.createTakeProfitDetails(1.2345, info)
}
