% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createTakeProfitOrder}
\alias{oanda.orders.createTakeProfitOrder}
\title{Create a take profit order}
\usage{
oanda.orders.createTakeProfitOrder(account, tradeID, price, clientExtensions)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{tradeID}{The ID of the Trade to close when the price threshold is breached.}

\item{price}{The price threshold specified for the Limit Order. The Limit Order will only be filled by a market price that is equal to or better than this price.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
R list object detailing the response from the server.
}
\description{
This function initiates take profit order. A TakeProfitOrder is an order that is linked
to an open Trade and created with a price threshold.
}
\examples{
oanda.orders.createTakeProfitOrder('101-001-3704066-001', "6368", 1.32821) 
info <- oanda.orders.createMarketIfTouchedOrder("A comment","strategy_9","my_order_10000")
oanda.orders.createTakeProfitOrder('101-001-3704066-001', "6368", 1.32821, clientExtensions = info) 
}
