% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createTrailingStopLossDetails}
\alias{oanda.orders.createTrailingStopLossDetails}
\title{Create Trailing Stop Loss details}
\usage{
oanda.orders.createTrailingStopLossDetails(distance, clientExtensions)
}
\arguments{
\item{distance}{Specifies the distance (in price units) from the Trade’s open price to use as the Stop Loss Order price. Only one of the distance and price fields may be specified.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
A JSON object of StopLossDetails.
}
\description{
TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be created on behalf of a client.
This may happen when an Order is filled that opens a Trade requiring a Trailing Stop Loss, or
when a Trade’s dependent Trailing Stop Loss Order is modified directly through the Trade.
}
\examples{
oanda.orders.createTrailingStopLossDetails(distance = 0.0500)
info <- oanda.orders.createClientExtensions("A comment","strategy_9","my_order_10000")
oanda.orders.createStopLossDetails(distance = 0.0100, info)
}
