% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.createTrailingStopLossOrder}
\alias{oanda.orders.createTrailingStopLossOrder}
\title{Create a trailing stop loss order}
\usage{
oanda.orders.createTrailingStopLossOrder(
  account,
  tradeID,
  distance,
  clientExtensions
)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{tradeID}{The ID of the Trade to close when the price threshold is breached.}

\item{distance}{Specifies the distance (in price units) from the Trade’s open price to use as the Stop Loss Order price. Only one of the distance and price fields may be specified.}

\item{clientExtensions}{The Client Extensions to add to the Take Profit Order when created. This is optional.}
}
\value{
R list object detailing the response from the server.
}
\description{
A TrailingStopLossOrderRequest specifies the parameters that may be set when creating a Trailing Stop Loss Order.
}
\examples{
info <- oanda.orders.createMarketIfTouchedOrder("A comment","strategy_9","my_order_10000")
oanda.orders.createTrailingStopLossOrder('101-001-3704066-001', "6368", distance = 0.02821)
oanda.orders.createTrailingStopLossOrder('101-001-3704066-001', "6368", distance = 0.02821, clientExtensions = info)
}
