% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.orders.pending}
\alias{oanda.orders.pending}
\title{Pending Orders}
\usage{
oanda.orders.pending(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
R list object detailing the response from the server.
}
\description{
List all pending orders in an account
}
\examples{
oanda.orders.pending('101-001-3704066-001')
}
