% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.positions.close}
\alias{oanda.positions.close}
\title{Close position by instrument.}
\usage{
oanda.positions.close(account, instrument, longUnits, shortUnits)
}
\arguments{
\item{account}{The string representation of an Account identifier.}

\item{instrument}{The name of the instrument.}

\item{longUnits}{Indication of how much of the long Position to closeout. Either the string “ALL”, the string “NONE”,
or a DecimalNumber representing how many units of the long position to close using a PositionCloseout MarketOrder.
he units specified must always be positive.}

\item{shortUnits}{Indication of how much of the short Position to closeout. Either the string “ALL”, the string “NONE”,
or a DecimalNumber representing how many units of the short position to close using a PositionCloseout MarketOrder.
The units specified must always be positive.}
}
\value{
R list object detailing the response from the server.
}
\description{
Close position by instrument.
}
\examples{
oanda.positions.close('101-001-3704066-001',"USD_NOK") #close all positions of USD/NOK
oanda.positions.close('101-001-3704066-001',"USD_NOK", shortUnits = 5000)
#close 5000 units of short position in USD/NOK
oanda.positions.close('101-001-3704066-001',"USD_SGD", longUnits = 5000)
#close 5000 units of long position in USD/SGD
}
