% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.positions.instrument}
\alias{oanda.positions.instrument}
\title{List position by instrument.}
\usage{
oanda.positions.instrument(account, instrument)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instrument}{The name of the instrument to query positions by.}
}
\value{
R list object detailing the response from the server.
}
\description{
List position by instrument.
}
\examples{
oanda.positions.instrument('101-001-3704066-001',"USD_MXN")
}
