% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.positions.open}
\alias{oanda.positions.open}
\title{List open positins}
\usage{
oanda.positions.open(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
R list object detailing the response from the server.
}
\description{
List open positins
}
\examples{
oanda.positions.open('101-001-3704066-001')
}
