% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.positions}
\alias{oanda.positions}
\title{List Positions}
\usage{
oanda.positions(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
R list object detailing the response from the server.
}
\description{
List Positions
}
\examples{
oanda.positions('101-001-3704066-001')
}
