% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.trades.byId}
\alias{oanda.trades.byId}
\title{Query open trade by ID}
\usage{
oanda.trades.byId(account, tradeId)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{tradeId}{The ID to filter the open trades by.}
}
\value{
R list object detailing the response from the server.
}
\description{
Filter open trades for an account by trade ID.
}
\examples{
oanda.trades.byId('101-001-3704066-001',"5878")
}
