% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.trades.byIntrument}
\alias{oanda.trades.byIntrument}
\title{Query open trades by Instrument}
\usage{
oanda.trades.byIntrument(account, instrument)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{instrument}{The name of the instrument to query by.}
}
\value{
R list object detailing the response from the server.
}
\description{
Filter open trades for an account by instrument.
}
\examples{
oanda.trades.byIntrument('101-001-3704066-001',"USD_SGD")
}
