% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.trades.closeById}
\alias{oanda.trades.closeById}
\title{Close a trade by ID}
\usage{
oanda.trades.closeById(account, tradeId, quantity)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{tradeId}{The trade ID.}

\item{quantity}{Indication of how much of the Trade to close. Either the string “ALL” (indicating that all of the Trade should be closed),
or a DecimalNumber representing the number of units of the open Trade to Close using a TradeClose MarketOrder}
}
\value{
R list object detailing the response from the server.
}
\description{
Close an open trade by ID
}
\examples{
oanda.trades.closeById('101-001-3704066-001',"6985") #close all trades
oanda.trades.closeById('101-001-3704066-001',"6988", 5000)
#close 5000 units assuming there are 10000 units in the trade
oanda.trades.closeById('101-001-3704066-001',"6987", 0.5) #will close half of the position held.
}
