% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.trades.editById}
\alias{oanda.trades.editById}
\title{Edit an open trade}
\usage{
oanda.trades.editById(account, tradeId, takeProfit, stopLoss, trailingStopLoss)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}

\item{tradeId}{The trade ID.}

\item{takeProfit}{The take profit price.}

\item{stopLoss}{The stop loss price.}

\item{trailingStopLoss}{The trailing stop loss distance.}
}
\value{
R list object detailing the response from the server.
}
\description{
Create, replace and cancel a Trade’s dependent Orders (Take Profit, Stop Loss and Trailing Stop Loss)
hrough the Trade itself
}
\examples{
oanda.trades.editById('101-001-3704066-001', "6985", takeProfit = 1.34567)
oanda.trades.editById('101-001-3704066-001', "6985", stopLoss = 1.23456)
oanda.trades.editById('101-001-3704066-001', "6985", distance = .00250)
}
