% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oanda.r
\name{oanda.trades.open}
\alias{oanda.trades.open}
\title{Open Trades}
\usage{
oanda.trades.open(account)
}
\arguments{
\item{account}{The string representation of an Account Identifier.}
}
\value{
R list object detailing the response from the server.
}
\description{
List all open trades for an account.
}
\examples{
oanda.trades.open('101-001-3704066-001')
}
