% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{pricing}
\alias{pricing}
\title{Get pricing}
\usage{
pricing(base_currency, quote_currency, since = Sys.time() - 3600,
  account_id = Sys.getenv("OANDA_ACCOUNT_ID", NA),
  access_token = Sys.getenv("OANDA_ACCESS_TOKEN", NA))
}
\arguments{
\item{base_currency}{A 3 letter ISO 4217 currency name.}

\item{quote_currency}{A 3 letter ISO 4217 currency name.}

\item{since}{A time in the past, or \code{NULL} for now.}
}
\value{
A \code{data_frame} with one row per currency pair.
}
\description{
Call the OANDA pricing API to retrieve prices of currency pairs.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
  (prices <- pricing(c("EUR", "GBP"), "USD"))
}
}
\references{
\url{http://developer.oanda.com/rest-live-v20/pricing-ep/#CurrentPrices}
}
